doModify, doDelete 권한체크

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -82,9 +82,11 @@ public class UsrArticleController {
 	public ResultData<Integer> doModify(HttpSession httpSession, int id, String title, String body) {
 
 		boolean isLogined = false;
+		int loginedMemberId = 0;
 
 		if (httpSession.getAttribute("loginedMemberId") != null) {
 			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
 		}
 
 		if (isLogined == false) {
@@ -95,6 +97,10 @@ public class UsrArticleController {
 
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다", id), id);
+		}
+
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", Ut.f("%d번 글에 대한 권한이 없습니다", id), id);
 		}
 
 		articleService.modifyArticle(id, title, body);
@@ -107,19 +113,24 @@ public class UsrArticleController {
 	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
 
 		boolean isLogined = false;
+		int loginedMemberId = 0;
 
 		if (httpSession.getAttribute("loginedMemberId") != null) {
 			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
 		}
 
 		if (isLogined == false) {
 			return ResultData.from("F-A", "로그인 후 이용해주세요");
 		}
-
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다", id), id);
+		}
+
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", Ut.f("%d번 글에 대한 권한이 없습니다", id), id);
 		}
 
 		articleService.deleteArticle(id);
